fix(houses): change calc base into utc to calc base into calendar day
Previously, this implementation was using the recommended methods for
dates based on Coordinated Universal Time (UTC). These functions are
recommended more for hours obtained from clock. The problem with them
is that for some calculations they deliver differences with other
libraries that use swesseph. Due to standardization issues with other
libraries and to avoid confusion, it was decided to replace the
implementation to be able to use the dates based on the calendar day.